Refactor LLM provider implementations to enhance error handling, support new message formats, and improve serialization/deserialization with serde. Added LLMClient trait for unified client interface across providers. Updated Gemini, Anthropic, LMStudio, and OpenAI providers to handle tool calls and system prompts more effectively. Improved response parsing and added debug logging for better traceability. Adjusted model handling and default values for requests.

+++ vtagent-core/src/core/agent/multi_agent.rs
@@ -578,7 +578,7 @@ impl Default for MultiAgentConfig {
             delegation_strategy: DelegationStrategy::Adaptive,
             context_store: ContextStoreConfig {
                 max_contexts: ContextStoreDefaults::MAX_CONTEXTS,
-                auto_cleanup_days: ContextStoreDefaults::AUTO_CLEANUP_DAYS,
+                auto_cleanup_days: ContextStoreDefaults::AUTO_CLEANUP_DAYS as u64,
                 enable_persistence: ContextStoreDefaults::ENABLE_PERSISTENCE,
                 compression_enabled: ContextStoreDefaults::COMPRESSION_ENABLED,
                 storage_dir: ContextStoreDefaults::STORAGE_DIR.to_string(),
